requirements.txt
opencv-python
numpy
pandas
tqdm
gdown
openslide-python
torchstain>=1.3.1
scikit-image
requests

# For β-VAE training
torch
torchvision
albumentations
pytorch-msssim
scikit-learn
scipy
matplotlib
Pillow

